+++ ImTK/UI/VisualTree/NodeType.cs
@@ -1,0 +1,10 @@
+namespace ImTK.UI
+{
+    public enum NodeType
+    {
+        None,
+        LogicNode,
+        PhysicsNode,
+        Invalid
+    }
+}